+++ packages/url/src/index.ts
@@ -1,3 +1,4 @@
+import { __count } from '@reatom/core'
 import { Action, Atom, AtomMut, AtomState, Ctx, Fn, Rec, action, atom } from '@reatom/core'
 import { abortCauseContext } from '@reatom/effects'
 import { getRootCause, withInit } from '@reatom/hooks'
@@ -30,6 +31,7 @@ export interface SearchParamsAtom extends Atom<Rec<string>> {
       parse?: (value?: string) => T
       serialize?: (value: T) => undefined | string
       replace?: boolean
+      path?: string
     },
   ): AtomMut<T>
 }
@@ -162,7 +164,7 @@ export const searchParamsAtom: SearchParamsAtom = Object.assign(
       urlAtom(ctx, newUrl, replace)
     }, 'searchParamsAtom._del') satisfies SearchParamsAtom['del'],
     lens: ((key, ...a: Parameters<typeof getSearchParamsOptions>) =>
-      atom(getSearchParamsOptions(...a).parse(), `searchParamsAtom#${a[0]}`).pipe(
+      atom(getSearchParamsOptions(...a).parse(), __count('searchParamsAtom')).pipe(
         // TODO
         // @ts-expect-error
         withSearchParamsPersist(key, ...a),
@@ -178,6 +180,7 @@ const getSearchParamsOptions = (
           parse?: (value?: string) => unknown
           serialize?: (value: unknown) => undefined | string
           replace?: boolean
+          path?: string
         },
       ]
 ) => {
@@ -185,20 +188,29 @@ const getSearchParamsOptions = (
     parse = (value = '') => String(value),
     serialize = (value: any) => (value === init ? undefined : String(value)),
     replace,
+    path,
   } = typeof a[0] === 'object'
     ? a[0]
     : {
         parse: a[0],
         serialize: a[1],
         replace: undefined,
+        path: undefined,
       }
   const init = parse()
   return {
     parse,
     serialize,
     replace,
+    path,
   }
 }
+
+const isSubpath = (currentPath: string, targetPath: string) =>
+  !targetPath ||
+  (targetPath[targetPath.length - 1] === '*'
+    ? `${currentPath}/`.startsWith(targetPath.slice(0, -1))
+    : `${currentPath}/` === targetPath)
 
 export function withSearchParamsPersist<T = string>(
   key: string,
@@ -211,6 +223,7 @@ export function withSearchParamsPersist<T = string>(
     parse?: (value?: string) => T
     serialize?: (value: T) => undefined | string
     replace?: boolean
+    path?: string
   },
 ): <A extends Atom<T>>(theAtom: A) => A
 export function withSearchParamsPersist(
@@ -220,22 +233,33 @@ export function withSearchParamsPersist(
     | [
         options: {
           parse?: (value?: string) => unknown
-          serialize?: (value: unknown) => undefined | string
+          serialize?: (value?: unknown) => undefined | string
           replace?: boolean
+          path?: string
         },
       ]
 ) {
-  const { parse, serialize, replace } = getSearchParamsOptions(...a)
+  let { parse, serialize, replace, path = '' } = getSearchParamsOptions(...a)
+
+  const pathEnd = path[path.length - 1]
+  if (path && pathEnd !== '/' && pathEnd !== '*') {
+    path += '/'
+  }
 
   return (theAtom: Atom) => {
     const { computer, initState } = theAtom.__reatom
+
     theAtom.pipe(
       withInit((ctx, init) => {
         const sp = ctx.get(searchParamsAtom)
-        return key in sp ? parse(sp[key]) : init(ctx)
+        const currentPath = ctx.get(urlAtom).pathname
+
+        return key in sp && isSubpath(currentPath, path) ? parse(sp[key]) : init(ctx)
       }),
     )
+
     theAtom.__reatom.computer = (ctx, state) => {
+      const currentPath = ctx.get(urlAtom).pathname
       ctx.spy(searchParamsAtom, (next, prev) => {
         if (key in next) {
           if (!prev || prev[key] !== next[key]) {
@@ -243,10 +267,17 @@ export function withSearchParamsPersist(
           }
         } else {
           if (prev && key in prev) {
+            if (isSubpath(currentPath, path)) {
+              const prevState = state as string
+              ctx.schedule(() => {
+                searchParamsAtom.set(ctx, key, prevState, true)
+              }, 0)
+            }
             state = initState(ctx)
           }
         }
       })
+
       if (computer) {
         const { pubs } = ctx.cause
 
@@ -274,14 +305,18 @@ export function withSearchParamsPersist(
           }
         }
       }
+
       return state
     }
+
     theAtom.onChange((ctx, state) => {
-      const value = serialize(state)
-      if (value === undefined) {
-        searchParamsAtom.del(ctx, key, replace)
-      } else {
-        searchParamsAtom.set(ctx, key, value, replace)
+      if (ctx.cause === theAtom.__reatom.patch) {
+        const value = serialize(state)
+        if (value === undefined) {
+          searchParamsAtom.del(ctx, key, replace)
+        } else if (isSubpath(ctx.get(urlAtom).pathname, path)) {
+          searchParamsAtom.set(ctx, key, value, replace)
+        }
       }
       ctx.get(theAtom)
     })